feat: separate implementations
