--- src/Uno.UI/UI/Xaml/Controls/FlipView/FlipView.wasm.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Linq;
-using System.Threading.Tasks;
-using Uno.Extensions;
-using Uno.UI.Extensions;
-using System.Collections.Specialized;
-using Uno.Extensions.Specialized;
-using Windows.UI.Xaml.Controls.Primitives;
-using Uno;
-
-namespace Windows.UI.Xaml.Controls
-{
-	public partial class FlipView : Selector
-	{
-
-	}
-}
-
